Safari-611.1.10/LayoutTests/http/tests/security/top-level-unique-origin2.https-expected.txt

CONSOLE MESSAGE: data URL frame loaded
CONSOLE MESSAGE: blob popup opened
CONSOLE MESSAGE: blob popup loaded
CONSOLE MESSAGE: blob popup message posted
PASS
